+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Add a browser-independent J2CL target backed by the deterministic test scheduler.
+* Use the global `console` object for J2CL logging.
+
 ### [v0.17](https://github.com/realityforge/zemeckis/tree/v0.17) (2026-07-24) · [Full Changelog](https://github.com/realityforge/zemeckis/compare/v0.16...v0.17)
 
 Changes in this release: